Refactors extf_mul, moves curve_id specifics to lib.rs
